hf_space/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
gradio==4.44.0
together==0.2.0
